+++ gamemodes/terrortown/gamemode/server/sv_player_ext.lua
@@ -1458,7 +1458,6 @@ local function SetPlayerReady(_, ply)
 	-- update playermodels on the client
 	if ply:IsSuperAdmin() then
 		playermodels.StreamModelStateToSelectedClients(ply)
-		playermodels.StreamModelStateToSelectedClients(ply, true)
 	end
 
 	---